Move resource permits into worker-api
