+++ crates/vqueues/src/scheduler/resource_manager.rs
@@ -15,12 +15,11 @@ mod locks;
 mod permit;
 mod user_limiter;
 
-pub use self::permit::{PermitBuilder, ReservedResources};
+pub use self::permit::PermitBuilder;
 
 use std::collections::VecDeque;
 use std::task::Poll;
 
-use smallvec::SmallVec;
 use tokio::sync::mpsc;
 use tracing::trace;
 
@@ -32,12 +31,13 @@ use restate_storage_api::vqueue_table::{EntryKey, EntryMetadata};
 use restate_types::vqueues::EntryKind;
 use restate_types::{LockName, Scope};
 use restate_worker_api::ResourceKind;
+use restate_worker_api::resources::{ResourceManagerUpdate, UserPermitKind};
 
 use self::invoker::InvokerConcurrencyLimiter;
 use self::invoker_memory::InvokerMemoryLimiter;
 use self::invoker_throttle::{InvokerThrottlingLimiter, ThrottlingAcquire};
 use self::locks::Locks;
-use self::permit::{ProvisionalPermit, UserPermitKind};
+use self::permit::ProvisionalPermit;
 use self::user_limiter::UserLimiter;
 use super::VQueueHandle;
 use super::eligible::EligibilityTracker;
@@ -59,12 +59,6 @@ pub struct ResourceManager {
     // We need to keep this alive to:
     // - Keep the receiver alive even if we don't have any resource permits handed out
     tx: mpsc::UnboundedSender<ResourceManagerUpdate>,
-}
-
-#[allow(dead_code)]
-enum ResourceManagerUpdate {
-    PermitReleased(SmallVec<[UserPermitKind; 1]>),
-    RulesUpdated(user_limiter::RuleUpdate),
 }
 
 pub(super) enum AcquireOutcome {
