+++ crates/worker-api/Cargo.toml
@@ -24,11 +24,12 @@ restate-storage-api = { workspace = true }
 restate-types = { workspace = true }
 restate-util-string = { workspace = true }
 
-codederror = { workspace = true }
-gardal = { workspace = true, features = ["tokio"] }
 bytes = { workspace = true }
+codederror = { workspace = true }
 futures = { workspace = true }
+gardal = { workspace = true, features = ["tokio"] }
 serde = { workspace = true, optional = true }
+smallvec = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
