+++ crates/worker-api/src/lib.rs
@@ -9,6 +9,7 @@
 // by the Apache License, Version 2.0.
 
 pub mod invoker;
+pub mod resources;
 
 mod leader_query;
 mod partition_processor_manager;
